mon penis yer mort

+++ BartEngine_1.0.0/bartengine/includes/Background.h
@@ -22,7 +22,7 @@ namespace bart
         void Unload();
 
     protected:
-        unsigned mTexId{0};
+        size_t mTexId{0};
         Rectangle m_DestinationA{};
         Rectangle m_DestinationB{};
         Rectangle m_SourceA{};
