+++ BartEngine_1.0.0/bartengine/includes/Sound.h
@@ -17,7 +17,7 @@ namespace bart
         void Unload();
 
     private:
-        unsigned m_SoundId{0};
+        size_t m_SoundId{0};
     };
 }
